add the thermometer image component to Storybook
