+++ src/components/Thermometer/index.ts
@@ -1,5 +1,6 @@
 import Thermometer from './Thermometer';
 import { ThermometerBox } from './Thermometer.style';
+import ThermometerImage from './ThermometerImage';
 
 export default Thermometer;
-export { ThermometerBox };
+export { ThermometerBox, ThermometerImage };
